refactor: Outsource `makers` from collectors
